+++ src/TestTrackingDiagrams.Extensions.CloudStorage/CloudStorageTrackingVerbosity.cs
@@ -1,0 +1,8 @@
+namespace TestTrackingDiagrams.Extensions.CloudStorage;
+
+public enum CloudStorageTrackingVerbosity
+{
+    Raw,
+    Detailed,
+    Summarised
+}